pyproject.toml
[tool.poetry]
name = "task-companion-ai"
version = "0.1.0"
description = "TaskCompanion is an AI-driven task management tool that helps users manage their tasks efficiently with AI-generated suggestions, reminders, and scheduling."
authors = ["Reuben John <reubenvjohn@gmail.com>"]
readme = "README.md"
packages = [{include = "main"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
langchain = "^0.1.12"
openai = "^1.14.1"
django-rest-framework = "^0.1.0"

[tool.poetry.scripts]
manage = "manage:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
